requirements.txt
fluent-logger>=0.9.3 # Apache-2.0
python-json-logger>=0.1.9 # BSD
grpcio>=1.13.0 # Apache-2.0
grpcio-tools>=1.13.0 # Apache-2.0
PyMySQL>=0.8.0 # MIT
SQLAlchemy>=1.2.7 # MIT
PyYAML>=3.12 # MIT
boto>=2.49.0 # MIT
